Refactor stack drag and drop into separate files
