+++ app/utils/shared/stack-utils.ts
@@ -1,0 +1,10 @@
+import { Stack } from '@prisma/client';
+
+export function recalcStackPositions(stacks: Stack[]): Stack[] {
+  return stacks.map((stack, index) => {
+    return {
+      ...stack,
+      position: index * 10,
+    };
+  });
+}
